requirements.txt
# --- Основной веб-фреймворк ---
fastapi
uvicorn
python-multipart

# --- HTTP-клиент для API Яндекса ---
httpx

# --- Обработка изображений ---
# rembg[gpu] для ускорения на NVIDIA GPU, используйте rembg для CPU
rembg[gpu]
Pillow

# --- Обработка аудио ---
pydub
soundfile

# --- AI & Voice Conversion ---
# Важно: torch нужно установить ОТДЕЛЬНОЙ командой до этого файла.
# Эта библиотека установит нужную версию fairseq как зависимость.
rvc-python